% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\name{.onLoad}
\alias{.onLoad}
\title{Add nctables type mapping to options interface}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{unused but retained for historical reasons}

\item{pkgname}{unused but retained for historical reasons}
}
\description{
Add nctables type mapping to options interface
}
\author{
Ben Fasoli
}
