% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\name{.onUnload}
\alias{.onUnload}
\title{Remove nctables type mapping from options interface}
\usage{
.onUnload(libname, pkgname)
}
\arguments{
\item{libname}{unused but retained for historical reasons}

\item{pkgname}{unused but retained for historical reasons}
}
\description{
Remove nctables type mapping from options interface
}
\author{
Ben Fasoli
}
