% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nct_create.r
\name{nct_create}
\alias{nct_create}
\title{Serializes data.frame to NetCDF file}
\usage{
nct_create(df, filename, dims, longnames = NA, units = NA)
}
\arguments{
\item{df}{data.frame to be written}

\item{filename}{character string specifying output file ending in \code{.nc}}

\item{dims}{character vector specifying at least one column as a dimension}

\item{longnames}{optional long names to include as dimension or variable
attributes, formatted as a named list or vector where keys match column
names in \code{df} and values specify the desired \code{long_name} NetCDF attribute}

\item{units}{optional units to include as dimension or variable attributes,
using the same format as \code{longnames}, except POSIXct columns will override
a user specified units field with "seconds since 1970-01-01T00:00:00Z" for
POSIX compliance}
}
\description{
Serializes data.frame to NetCDF file
}
\author{
Ben Fasoli
}
