% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nct_deserialize.r
\name{nct_deserialize}
\alias{nct_deserialize}
\title{Deserialize object from nctables format}
\usage{
nct_deserialize(nc, var)
}
\arguments{
\item{nc}{an object of class \code{ncdf4} as returned by \code{nc_open}}

\item{var}{string indicating variable to deserialize}
}
\description{
Deserialize object from nctables format
}
\author{
Ben Fasoli
}
