% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nct_read.r
\name{nct_read}
\alias{nct_read}
\title{Reads nct formatted NetCDF file to a data.frame}
\usage{
nct_read(filename)
}
\arguments{
\item{filename}{character string specifying output file ending in \code{.nc}}
}
\description{
Reads nct formatted NetCDF file to a data.frame
}
\author{
Ben Fasoli
}
