% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nct_serialize.r
\name{nct_serialize}
\alias{nct_serialize}
\title{Coerce object into nctables compatible format}
\usage{
nct_serialize(x)
}
\arguments{
\item{x}{object to serialize, typically a column extracted from a data.frame}
}
\description{
Coerce object into nctables compatible format
}
\author{
Ben Fasoli
}
